requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
nltk==3.2.5
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SpeechRecognition==3.7.1
textblob==0.13.0
tweepy==3.5.0
urllib3==1.22
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
nltk==3.2.5
numpy==1.13.3
oauthlib==2.0.6
olefile==0.44
pandas==0.20.3
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SpeechRecognition==3.7.1
textblob==0.13.0
tweepy==3.5.0
urllib3==1.22
